==> R/clean_data/03_catch_curves_data_prep.Rmd <==
---
title: "Prepare data for catch curves"
author: "Max Lindmark"
date: "`r format(Sys.time(), '%d %B, %Y')`"
output:
  html_document:
    df_print: paged
  pdf_document: default
editor_options: 
  chunk_output_type: console
  markdown: 
    wrap: 72
---

```{r setup, include = FALSE, cache=FALSE}
knitr::opts_chunk$set(
  collapse = TRUE,
  comment = "#>",
  fig.width = 12,
  fig.asp = 0.618,
  fig.align ='center'
)
```

### Load libraries
```{r libraries and settings, message=FALSE}
# Load libraries (install first if needed)
library(tidyverse); theme_set(theme_classic(base_size = 12))
library(brms)
library(nlstools)
library(viridis)
library(bayesplot)
library(tidylog)
library(tidybayes)
library(RColorBrewer)
library(patchwork)
library(modelr)

# Print package versions
sessionInfo()
```

### Read and add final touches to data
```{r read and prep data for analysis}
df <- read.csv("data/cleaned/aged_catch_BT_FM_1987-2003.csv")

head(df)

# 1996 is a special year because stations changed. But since we don't do anything with station 
# we simply ignore that
# df %>%
#   group_by(year, Area, Station) %>% 
#   summarise(n = n()) %>% 
#   ggplot(., aes(year, n, fill = factor(Station))) +
#   geom_bar(stat = "identity") +
#   facet_wrap(~Area)

# df %>%
#   filter(Area == "BT") %>%
#   group_by(year, netID) %>%
#   summarise(n = n()) %>%
#   ggplot(., aes(year, n)) + geom_bar(stat="identity")

# df %>% group_by(Area) %>% summarise(mean_age = mean(age), sd_age = sd(age))
# df %>% filter(age > 2) %>% group_by(Area) %>% summarise(mean_age = mean(age), sd_age = sd(age))
# ggplot(df, aes(Area, age)) + geom_boxplot()
# df %>% filter(age > 2) %>% ggplot(., aes(Area, age)) + geom_boxplot()
# summary(lm(age~Area, data = df))
# summary(lm(age~Area, data = filter(df, age > 2)))

# Currently the data is in format 1 row, 1 ind. We want a column that has the CPUE
# by length class.


# How many nets in total per area and year? (For scaling with effort later)
df <- df %>% group_by(Area, year) %>%
  mutate(n_nets_year_test = length(unique(netID))) %>%
  mutate(n_nets_year = length(unique(netID2))) %>%
  ungroup() %>%
  as.data.frame()

length(unique(filter(df, Area == "BT")$netID))
length(unique(filter(df, Area == "FM")$netID))

length(unique(filter(df, Area == "BT")$netID2))
length(unique(filter(df, Area == "FM")$netID2))

# We want the data as follows:
# Year, LngtClass, Age, Number

# Group by year, Area and age, summarize and get n() per length, year and area...
df2 <- df %>%
  group_by(year, Area, age) %>% 
  summarise(catch_n = n()) %>% # Get catch as a column, this is now the number of rows a given length occurs each year and area
  ungroup() %>% 
  mutate(effort_id = paste(year, Area, sep = ".")) %>% # For calculating the total effort that gave that catch (total for year and area) 
  as.data.frame()

# Now we need to get the effort back in there
df_effort <- df %>%
  mutate(effort_id = paste(year, Area, sep = ".")) %>% 
  select(effort_id, n_nets_year_test, n_nets_year) %>% 
  distinct(effort_id, .keep_all = TRUE) %>% 
  as.data.frame()

# Now do a left_join
df3 <- left_join(df2, df_effort, by = "effort_id") %>% as.data.frame()

# Looks Ok!
df %>% filter(year == 1987 & Area == "FM")
df %>% filter(year == 1987 & Area == "FM") %>% distinct(n_nets_year, .keep_all = TRUE)
df2 %>% filter(effort_id == "1987.FM")
df3 %>% filter(effort_id == "1987.FM")

# Go from total catch to CPUE
df4 <- df3 %>% 
  ungroup() %>% 
  rename("area" = "Area") %>% 
  mutate(cpue_numbers_test = catch_n/n_nets_year_test,
         cpue_numbers = catch_n/n_nets_year) # Get numbers CPUE, divide by the previously create n_nets, which is # of unique net ID's in each area and year
   
head(df3, 30)
head(df4)

p1 <- df4 %>% 
  group_by(area, year) %>% 
  summarise(sum_cpue = sum(cpue_numbers)) %>% 
  ggplot(aes(year, sum_cpue, color = area)) +
  geom_point() + 
  stat_smooth() + 
  ggtitle("long ID")

p2 <- df4 %>% 
  group_by(area, year) %>% 
  summarise(sum_cpue = sum(cpue_numbers_test)) %>% 
  ggplot(aes(year, sum_cpue, color = area)) +
  geom_point() + 
  stat_smooth() +
  ggtitle("short ID")

p1/p2
      
# Calculate average cpue by year and area
df4

p1 <- df4 %>%
  group_by(year, area, age) %>% 
  summarise(mean_cpue = mean(cpue_numbers)) %>% 
  ggplot(aes(year, mean_cpue, fill = factor(age))) +
  geom_area() + 
  facet_wrap(~area) + 
  scale_fill_brewer(palette = "Set1") + 
  coord_cartesian(expand = 0) + 
  ggtitle("New ID")

p2 <- df4 %>%
  group_by(year, area, age) %>% 
  summarise(mean_cpue = mean(cpue_numbers_test)) %>% 
  ggplot(aes(year, mean_cpue, fill = factor(age))) +
  geom_area() + 
  facet_wrap(~area) + 
  scale_fill_brewer(palette = "Set1") + 
  coord_cartesian(expand = 0) +
  ggtitle("Old ID")

p1/p2

# Test I get 1 unique row per age, year and area
df4 %>%
  group_by(year, area, age) %>%
  summarise(n = n()) %>% 
  ungroup() %>% 
  distinct(n)

# Plot the log cpue as function of age to find the ages that correspond to the descending limb:
df4 %>% 
  mutate(area2 = ifelse(area == "BT", "Heated", "Ref")) %>% 
  ggplot(aes(factor(age), log(cpue_numbers), color = factor(year))) + 
  geom_jitter(height = 0, width = 0.2, alpha = 0.8) +
  facet_wrap(~ area2) +
  scale_color_viridis(discrete = TRUE, name = "Year") + 
  labs(y = "Catch per unit effort", x = "Age") +   
  theme(aspect.ratio = 1, 
        legend.position = "bottom")

ggsave("figures/supp/descending_limb.pdf", width = 20, height = 20, unit = "cm")  

# Let's use age 3 and older
ggplot(df4, aes(factor(age), log(cpue_numbers), color = factor(year))) + 
  geom_point() +
  facet_wrap(~ area) +
  scale_color_viridis(discrete = TRUE) + 
  coord_cartesian(expand = 0) + 
  theme(aspect.ratio = 1)

# Plot total catch
df2 %>%
  group_by(Area, year) %>%
  summarize(tot_catch = sum(catch_n)) %>% 
  ggplot(., aes(year, tot_catch)) + geom_bar(stat = "identity") +
  facet_wrap(~ Area)

# Plot cpue
df4 %>%
  group_by(area, year) %>%
  ggplot(., aes(year, cpue_numbers)) + geom_bar(stat = "identity") +
  facet_wrap(~ area)

# Edit variables
d <- df4 %>%
  mutate(log_cpue = log(cpue_numbers),
         area2 = ifelse(area == "BT", "Warm", "Cold"),
         cohort = year - age) 

# Remove cohorts before 1981
d <- d %>% filter(cohort > 1980, age > 2)

sort(unique(d$cohort))

max(d$year)
```

### Explore data
```{r exploratory plots}
# Plot data
d %>% filter(area == "BT") %>% 
  ggplot(aes(age, log_cpue)) + 
  geom_point() + 
  stat_smooth(method = "lm", se = FALSE) + 
  facet_wrap(~ cohort)

d %>% filter(area == "FM") %>% 
  ggplot(aes(age, log_cpue)) + 
  geom_point() + 
  stat_smooth(method = "lm", se = FALSE) + 
  facet_wrap(~ cohort)

d %>% 
  ggplot(aes(age, cpue_numbers)) + 
  geom_point() + 
  stat_smooth(method = "lm", se = FALSE) + 
  facet_wrap(~ area)

# Make age proportion data frame (not by year)
tot_cpue_area <- d %>%
  filter(age > 2) %>% 
  mutate(Area2 = ifelse(area == "BT", "Warm", "Cold")) %>% 
  group_by(Area2) %>% 
  summarise(sum_cpue = sum(cpue_numbers))

d_prop_age <- d %>%
  filter(age > 2) %>% 
  mutate(Area2 = ifelse(area == "BT", "Warm", "Cold")) %>% 
  group_by(Area2) %>% 
  mutate(sum_cpue = sum(cpue_numbers)) %>% 
  ungroup() %>% 
  group_by(Area2, age) %>% 
  summarise(sum_cpue_age = sum(cpue_numbers)) %>% 
  ungroup() 

d_prop_age %>% as.data.frame()

d_prop_age <- left_join(d_prop_age, tot_cpue_area) %>% 
  mutate(prop = sum_cpue_age / sum_cpue)

pal <- rev(brewer.pal(n = 6, name = "Paired")[c(2, 6)])

p <- d_prop_age %>% 
  ggplot(aes(factor(age), prop, color = Area2)) +
  geom_point(stat = "identity") +
  scale_color_manual(values = rev(pal)) +
  theme_light() + 
  NULL

# Plot proportion by age

pp <- df %>% 
  filter(age > 2) %>% 
  mutate(Area2 = ifelse(Area == "BT", "Warm", "Cold")) %>% 
  group_by(Area2, age) %>% 
  summarise(n = n()) %>%
  mutate(freq = n / sum(n)) %>% 
  ggplot(aes(x = factor(age), y = freq, color = Area2, group = interaction(age, Area2))) +
  geom_point() +
  scale_color_manual(values = rev(pal)) +
  theme_light() + 
  labs(x = "Age", y = "Proportion") +
  guides(fill = "none") +
  theme(text = element_text(size = 12))

p / pp

df %>% 
  filter(age > 2) %>% 
  mutate(Area2 = ifelse(Area == "BT", "Warm", "Cold")) %>% 
  group_by(Area2, year, age) %>% 
  summarise(n = n()) %>%
  mutate(freq = n / sum(n)) %>% 
  ggplot(aes(x = factor(age), y = freq, color = Area2, group = interaction(age, Area2))) +
  geom_point(position = position_dodge(width = 0.8)) +
  geom_boxplot(position = position_dodge(width = 0.8), fill = NA) +
  scale_color_manual(values = rev(pal)) +
  theme_light() + 
  labs(x = "Age", y = "Proportion") +
  guides(fill = "none") +
  theme(text = element_text(size = 12))

#ggsave("figures/supp/age_prop_diff.png", width = 3.5, height = 6.5, dpi = 600)
#ggsave("figures/supp/age_prop_diff.pdf", width = 3.5, height = 6.5, dpi = 600)
```

### Save data
```{r save data}
write.csv(d, "data/for_fitting/catch_curve_dat.csv")
```
